moved display items logic into create to allow for auto refresh
