+++ src/components/items/AddStockPage.jsx
@@ -1,15 +1,13 @@
 import CreateItem from "./CreateItem";
-import DisplayItems from "./DisplayItems";
 
 function AddStockPage() {
     return (
         <div>
             <CreateItem />
-            <DisplayItems />
 
         </div>
-        
-        );
+
+    );
 }
 
 export default AddStockPage;
