+++ src/components/items/CreateItem.jsx
@@ -1,5 +1,8 @@
 import axios from "axios";
-import { useState } from "react";
+import { useEffect, useState } from "react";
+import { Card } from "react-bootstrap";
+import { useNavigate } from 'react-router-dom';
+import Logo from "./Logo.png"
 
 
 function CreateItem() {
@@ -8,8 +11,21 @@ function CreateItem() {
   const [price, setPrice] = useState("");
   const [quantity, setQuantity] = useState("");
   const [image, setImage] = useState("");
-                     
-  
+  const [items, setItems] = useState([]);
+  const navigate = useNavigate();
+
+
+  function getItems() {
+    axios.get("http://localhost:8081/item/get")
+      .then(response => {
+        setItems(response.data);
+        console.log("http://localhost:8081/item/get", response);
+      })
+      .catch(err => console.error(err));
+  }
+
+  useEffect(() => getItems());
+
   const handleSubmit = (event) => {
     event.preventDefault();
 
@@ -28,6 +44,8 @@ function CreateItem() {
         setPrice("");
         setQuantity("");
         setImage("");
+        getItems();
+
 
       })
       .catch((error) => console.error(error));
@@ -58,10 +76,33 @@ function CreateItem() {
         />
         <br />
         <button type="submit" className="btn btn-success btn-md" style={{ backgroundColor: "#1C2938", borderColor: "white", color: "white" }}
->Create!</button>
+        >Create!</button>
 
       </form>
+      <br />
+      <br />
+      <div className="container">
+        {items.map(item => (
+          <Card key={item.id}
+            className="col-sm-6 col-md-4 col-lg-3 m-4"
+            style={{ textAlign: "center" }}>
+            <div style={{ position: "absolute", top: 10, right: 10, transform: "scale(2)" }}>
+              <img src={Logo} alt="Watermark" style={{ opacity: 0.65, width: 50 }} />
+            </div>
+            <img src={item.image} className="item-images" alt="itemImage" />
 
+            <br />
+            <h4>{item.name}</h4>
+            <h6> £ {item.price}</h6>
+            Items left: {item.quantity}
+            <br />
+            <br />
+
+            <button onClick={() => navigate("/edit/" + item.id)}>Edit Item</button>
+
+          </Card>
+        ))}
+      </div>
     </div>
   );
 }
